+++ sii/utils.py
@@ -1,5 +1,19 @@
 # -*- coding: UTF-8 -*-
 
+from unidecode import unidecode
+
+
+def unidecode_str(s):
+    if type(s) is unicode:
+        res = unidecode(s)
+    elif type(s) is str:
+        res = unidecode(s.decode('utf-8'))
+    else:
+        res = unidecode(s)
+
+    return res
+
+
 COUNTRY_CODES = {
     'AF': 'AFGANISTÁN',
     'AL': 'ALBANIA',
